requirements.txt
chromadb==1.5.5
ollama==0.6.1
pypdf==6.9.1
streamlit==1.55.0
numpy==2.4.3
python-dotenv==1.2.2
pytest==8.3.5
python-docx==1.1.2
